move WorldReadableRoomView out to its own file and create WorldReadableRoomViewModel right in the RoomViewModelObservable
create a new RoomStatus enum value for WorldReadable and utilise a promise to check whether room is previewable which first UnknownRoomViewModel uses to show UI context and then switches to WorldReadableRoomViewModel to render timeline for preview of messages
