+++ src/matrix/room/common.ts
@@ -35,6 +35,7 @@ export enum RoomStatus {
     Joined = 1 << 3,
     Replaced = 1 << 4,
     Archived = 1 << 5,
+    WorldReadable = 1 << 6,
 }
 
 export enum RoomType {
