+++ src/platform/web/ui/session/room/UnknownRoomView.js
@@ -14,21 +14,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import {InlineTemplateView, TemplateView} from "../../general/TemplateView";
-import {AvatarView} from "../../AvatarView";
-import {TimelineView} from "./TimelineView";
-import {TimelineLoadingView} from "./TimelineLoadingView";
+import {TemplateView} from "../../general/TemplateView";
 import {spinner} from "../../common.js";
-import {viewClassForTile} from "./common";
 
 export class UnknownRoomView extends TemplateView {
-
-    constructor(vm) {
-        super(vm);
-    }
-
     render(t, vm) {
-        return t.div({className: "UnknownRoomView middle"}, [
+        return t.main({className: "UnknownRoomView middle"}, t.div([
             t.h2([
                 vm.i18n`You are currently not in ${vm.roomIdOrAlias}.`,
                 t.br(),
@@ -45,41 +36,6 @@ export class UnknownRoomView extends TemplateView {
                 t.p(vm.i18n`Checking preview capability...`)
             ])),
             t.if(vm => vm.error, t => t.p({className: "error"}, vm.error))
-        ]);
-    }
-}
-
-export class WorldReadableRoomView extends TemplateView {
-
-    constructor(vm) {
-        super(vm);
-    }
-
-    render(t, vm) {
-        return t.div({className: "RoomView WorldReadableRoomView middle"}, [
-            t.div({className: "RoomHeader middle-header"}, [
-                t.view(new AvatarView(vm, 32)),
-                t.div({className: "room-description"}, [
-                    t.h2(vm => vm.room.name),
-                ]),
-            ]),
-            t.div({className: "RoomView_body"}, [
-                t.div({className: "RoomView_error"}, [
-                    t.if(vm => vm.error, t => t.div([
-                        t.p({}, vm => vm.error),
-                        t.button({className: "RoomView_error_closerButton", onClick: evt => vm.dismissError(evt)})
-                    ]))
-                ]),
-                t.mapView(vm => vm.timelineViewModel, timelineViewModel => {
-                    return timelineViewModel ?
-                        new TimelineView(timelineViewModel, viewClassForTile) :
-                        new TimelineLoadingView(vm);    // vm is just needed for i18n
-                }),
-                t.div({className: "WorldReadableRoomComposerView"}, [
-                    t.h3(vm => vm.i18n`Join the room to participate`),
-                    t.button({className: "joinRoomButton", onClick: () => vm.join()}, vm.i18n`Join Room`)
-                ])
-            ])
-        ]);
+        ]));
     }
 }
